% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{archer_fit7_odeint}
\alias{archer_fit7_odeint}
\title{Main optimization function for 8-parameter model}
\usage{
archer_fit7_odeint(parms, data)
}
\arguments{
\item{parms}{A numeric vector of parameter values.}

\item{data}{A dataframe containing necessary data (columns
\code{"Circ"} and \code{"ring_prop"}).}
}
\description{
Main optimization function for 8-parameter model
}
