% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizing.R
\name{winnowing_optim}
\alias{winnowing_optim}
\title{Winnowing optimization}
\usage{
winnowing_optim(
  fn,
  lower_bounds,
  upper_bounds,
  fn_args = list(),
  box_control = list(maxit = 100, reltol = 1e-04),
  fine_control = list(maxit = 500, reltol = 1e-06),
  polished_control = list(maxit = 1000, reltol = 1e-08),
  box_optim = optim,
  fine_optim = optim,
  polished_optim = optim,
  n_bevals = 100L,
  n_boxes = 1000L,
  n_fine = 100L,
  n_polished = 20L,
  n_outputs = 3L,
  file_bevals = NULL,
  file_boxes = NULL,
  file_fine = NULL,
  overwrite = FALSE,
  na_stop = FALSE
)
}
\arguments{
\item{fn}{Objective function to minimize.}

\item{lower_bounds}{Lower bounds of boxes for each parameter.
If this vector has names, these names are used for column names in
intermediate output tables (arguments starting with \code{file_} below).
If names are present, they must match those in \code{upper_bounds},
but don't need to be in the same order.}

\item{upper_bounds}{Upper bounds of boxes for each parameter.
See \code{lower_bounds} for providing names for this vector.}

\item{fn_args}{List containing other arguments to use for \code{fn}.
Defaults to \code{list()}.}

\item{box_control}{List containing arguments to use for the \code{control}
argument for the optimization that occurs for each box.
Defaults to \code{list(maxit = 100, reltol = 1e-4)}
(these names are adjusted if using an optimizer other than \code{optim}).}

\item{fine_control}{List containing arguments to use for the \code{control}
argument for the fine optimizations.
Defaults to \code{list(maxit = 500, reltol = 1e-6)}
(these names are adjusted if using an optimizer other than \code{optim}).}

\item{polished_control}{List containing arguments to use for the \code{control}
argument for the polished optimizations.
Defaults to \code{list(maxit = 1000, reltol = 1e-8)}
(these names are adjusted if using an optimizer other than \code{optim}).}

\item{box_optim}{Function to use for the optimization for each box.
Must be \code{stats::optim} or a function from packages \code{minqa} or \code{nloptr}.
Defaults to \code{stats::optim}.}

\item{fine_optim}{Function to use for the fine optimizations.
Must be \code{stats::optim} or a function from packages \code{minqa} or \code{nloptr}.
Defaults to \code{stats::optim}.}

\item{polished_optim}{Function to use for the polished optimizations.
Must be \code{stats::optim} or a function from packages \code{minqa} or \code{nloptr}.
Defaults to \code{stats::optim}.}

\item{n_bevals}{Number of evaluations of \code{fn} per box. Defaults to \code{100L}.}

\item{n_boxes}{Number of boxes. Defaults to \code{1000L}.}

\item{n_fine}{Number of fine optimizations.
Must be \verb{< n_boxes} and \verb{> n_polished}.
Defaults to \code{100L}.}

\item{n_polished}{Number of polished optimizations.
Must be \verb{< n_fine}.
Defaults to \code{20L}.}

\item{n_outputs}{Number of top output object(s) to return.
Must be \verb{< n_polished}. Defaults to \code{3L}.}

\item{file_bevals}{Single string specifying the filename where to save the
output from the box evaluations step.
The output is a table with the box number (column \code{box}),
parameter values (named based on bounds arguments or \code{par1}, \code{par2}, ...),
and output from objective function at those values (\code{val}).
If the file ends with \code{.txt}, the output will be tab-delimited,
and if the file ends with \code{.csv}, the output will be comma-delimited.
Other extensions are not allowed.
An error will trigger if attempting to overwrite an existing file
unless the \code{overwrite} argument is set to \code{TRUE}.
If \code{NULL} (the default), no output is written.}

\item{file_boxes}{Single string specifying the filename where to save the
output from the box evaluations step.
The output is a table with the starting parameter values (with \verb{_start} suffix),
ending parameter values (with \verb{_end} suffix),
convergence code for optimization (\code{conv_code}), and
output from objective function ending parameter values (\code{val}).
See description of argument \code{file_bevals} above for file extensions and
overwriting.
If \code{NULL} (the default), no output is written.}

\item{file_fine}{Single string specifying the filename where to save the
output from the box evaluations step.
The output is a table with the same columns as for \code{file_boxes}
(see above).
See description of argument \code{file_bevals} above for file extensions and
overwriting.
If \code{NULL} (the default), no output is written.}

\item{overwrite}{A single logical for whether to allow overwriting files
for intermediate output (arguments to this function starting
with \code{file_}). Defaults to \code{FALSE}.}

\item{na_stop}{Single logical for whether to return the matrix of initial
evaluations in each box if it contains \code{NA}s.
If \code{FALSE}, the optimizer ignores these values and continues on
(unless they're all \code{NA}).
Defaults to \code{FALSE}.}
}
\value{
A list containing \code{n_outputs} object(s) of the class returned
by \code{polished_optim}.
}
\description{
Winnowing optimization
}
